requirements.txt
absl-py==1.2.0
affine==2.3.1
aiohttp==3.8.3
aiosignal==1.2.0
albumentations==1.1.0
altair==4.2.0
anyio==3.6.1
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asttokens==2.0.5
astunparse==1.6.3
async-timeout==4.0.2
attrs==21.4.0
Babel==2.10.3
backcall==0.2.0
beautifulsoup4==4.11.1
bleach==5.0.1
blinker==1.5
cachetools==5.2.0
certifi==2022.5.18.1
cffi==1.15.1
charset-normalizer==2.0.12
click==8.1.3
click-plugins==1.1.1
cligj==0.7.2
cmdstanpy==1.0.7
colorama==0.4.4
commonmark==0.9.1
convertdate==2.4.0
cycler==0.11.0
Cython==0.29.32
debugpy==1.6.2
decorator==5.1.1
defusedxml==0.7.1
discord==2.0.0
discord.py==2.0.1
einops==0.4.1
entmax==1.0
entrypoints==0.4
ephem==4.1.3
executing==0.8.3
fastapi==0.86.0
fastjsonschema==2.16.1
filelock==3.7.1
Flask==2.2.1
Flask-Login==0.6.2
Flask-SQLAlchemy==2.5.1
flatbuffers==1.12
fonttools==4.37.1
frozenlist==1.3.1
gast==0.4.0
gitdb==4.0.9
GitPython==3.1.27
google-auth==2.11.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
greenlet==1.1.2
grpcio==1.48.1
gunicorn==20.1.0
h11==0.14.0
h5py==3.7.0
hijri-converter==2.2.4
holidays==0.15
huggingface-hub==0.7.0
idna==3.3
imageio==2.19.3
importlib-metadata==4.12.0
ipykernel==6.15.1
ipython==8.4.0
ipython-genutils==0.2.0
ipywidgets==7.7.1
itsdangerous==2.1.2
jedi==0.18.1
Jinja2==3.1.2
joblib==1.1.0
json5==0.9.8
jsonschema==4.7.2
jupyter==1.0.0
jupyter-client==7.3.4
jupyter-console==6.4.4
jupyter-core==4.11.1
jupyter-server==1.18.1
jupyterlab==3.4.4
jupyterlab-pygments==0.2.2
jupyterlab-server==2.15.0
jupyterlab-widgets==1.1.1
keras==2.9.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.4
korean-lunar-calendar==0.2.1
libclang==14.0.6
LunarCalendar==0.0.9
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.5.3
matplotlib-inline==0.1.3
mistune==0.8.4
multidict==6.0.2
munch==2.5.0
nbclassic==0.4.3
nbclient==0.6.6
nbconvert==6.5.0
nbformat==5.4.0
nest-asyncio==1.5.5
networkx==2.8.2
notebook==6.4.12
notebook-shim==0.1.0
numpy==1.22.4
oauthlib==3.2.0
opencv-python==4.6.0.66
opencv-python-headless==4.5.5.64
opt-einsum==3.3.0
ortools==9.3.10497
packaging==21.3
pandas==1.4.2
pandocfilters==1.5.0
parso==0.8.3
pickleshare==0.7.5
Pillow==9.1.1
pix2tex==0.0.26
plotly==5.10.0
prometheus-client==0.14.1
prompt-toolkit==3.0.30
prophet==1.1
protobuf==3.19.4
psutil==5.9.1
pure-eval==0.2.2
pyarrow==9.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pydantic==1.10.2
pydeck==0.8.0b1
Pygments==2.12.0
PyMeeus==0.5.11
Pympler==1.0.1
pynput==1.7.6
pyparsing==3.0.9
PyQt5==5.15.6
PyQt5-Qt5==5.15.2
PyQt5-sip==12.10.1
PyQtWebEngine==5.15.5
PyQtWebEngine-Qt5==5.15.2
pyrsistent==0.18.1
python-dateutil==2.8.2
python-multipart==0.0.5
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
PyWavelets==1.3.0
pywin32==304
pywinpty==2.0.6
PyYAML==6.0
pyzmq==23.2.0
qtconsole==5.3.1
QtPy==2.1.0
qudida==0.0.4
rasterio==1.3.2
regex==2022.6.2
requests==2.27.1
requests-oauthlib==1.3.1
rich==12.5.1
rsa==4.9
scikit-image==0.19.2
scikit-learn==1.1.1
scipy==1.8.1
screeninfo==0.8
semver==2.13.0
Send2Trash==1.8.0
setuptools-git==1.2
Shapely==1.8.4
six==1.16.0
smmap==5.0.0
sniffio==1.2.0
snuggs==1.4.7
soupsieve==2.3.2.post1
SQLAlchemy==1.4.39
stack-data==0.3.0
starlette==0.20.4
streamlit==1.12.2
tenacity==8.0.1
tensorboard==2.9.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.9.1
tensorflow-estimator==2.9.0
tensorflow-io-gcs-filesystem==0.26.0
termcolor==1.1.0
terminado==0.15.0
threadpoolctl==3.1.0
tifffile==2022.5.4
timm==0.5.4
tinycss2==1.1.1
tokenizers==0.12.1
toml==0.10.2
toolz==0.12.0
torch==1.11.0+cu113
torchaudio==0.11.0+cu113
torchvision==0.12.0+cu113
tornado==6.2
tqdm==4.64.0
traitlets==5.3.0
transformers==4.19.2
typing_extensions==4.2.0
tzdata==2022.2
tzlocal==4.2
ujson==5.4.0
urllib3==1.26.9
uvicorn==0.19.0
validators==0.20.0
watchdog==2.1.9
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.3.3
Werkzeug==2.2.1
widgetsnbextension==3.6.1
wrapt==1.14.1
x-transformers==0.15.0
yarl==1.8.1
zipp==3.8.1
